+++ pyvisim/encoders/fisher_vector.py
@@ -136,7 +136,7 @@ class FisherVectorEncoder(ClusteringBasedEncoder):
             )
         super()._set_clustering_model(clustering_model)
 
-    def encode(
+    def embed(
         self,
         images: ImageInput,
         *,